+++ example/ios/AirshipExampleNotificationServiceExtension/NotificationService.swift
@@ -1,0 +1,6 @@
+/* Copyright Airship and Contributors */
+
+import UserNotifications
+import AirshipServiceExtension
+
+class NotificationService: UANotificationServiceExtension {}