+++ example/lib/screens/named_user_add.dart
@@ -1,58 +1,51 @@
 import 'package:flutter/material.dart';
 import 'package:airship_example/styles.dart';
 import 'package:airship_example/widgets/text_add_bar.dart';
+// ignore: depend_on_referenced_packages
 import 'package:airship_flutter/airship_flutter.dart';
 
 class NamedUserAdd extends StatefulWidget {
-  final updateParent;
+  final VoidCallback updateParent;
 
-  NamedUserAdd({this.updateParent});
+  const NamedUserAdd({Key? key, required this.updateParent}) : super(key: key);
 
   @override
-  _NamedUserAddState createState() =>
-      _NamedUserAddState(updateParent: updateParent);
+  NamedUserAddState createState() => NamedUserAddState();
 }
 
-class _NamedUserAddState extends State<NamedUserAdd> {
-  final updateParent;
-
-  _NamedUserAddState({this.updateParent});
-
+class NamedUserAddState extends State<NamedUserAdd> {
   @override
   void initState() {
-    Airship.analytics.trackScreen('Add Named User');
     super.initState();
+    Airship.analytics.trackScreen('Add Named User');
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: Text("Add Named User"),
-          backgroundColor: Styles.background,
-        ),
-        body: FutureBuilder<String?>(
-          future: Airship.contact.namedUserId,
-          builder: (context, snapshot) {
-            return SafeArea(
-              bottom: false,
-              child: Column(
-                children: <Widget>[
-                  Padding(
-                    padding: const EdgeInsets.all(12.0),
-                    child: TextAddBar(
-                      label: snapshot.hasData ? snapshot.data! : "Not set",
-                      onTap: (text) {
-                        Airship.contact.identify(text);
-                        updateParent();
-                        Navigator.pop(context);
-                      },
-                    ),
-                  ),
-                ],
+      appBar: AppBar(
+        title: const Text("Add Named User"),
+        backgroundColor: Styles.background,
+      ),
+      body: FutureBuilder<String?>(
+        future: Airship.contact.namedUserId,
+        builder: (context, snapshot) {
+          return SafeArea(
+            bottom: false,
+            child: Padding(
+              padding: const EdgeInsets.all(12.0),
+              child: TextAddBar(
+                label: snapshot.data ?? "Not set",
+                onTap: (text) {
+                  Airship.contact.identify(text);
+                  widget.updateParent();
+                  Navigator.pop(context);
+                },
               ),
-            );
-          },
-        ));
+            ),
+          );
+        },
+      ),
+    );
   }
 }