+++ ios/Classes/AirshipPluginLoader.swift
@@ -1,0 +1,10 @@
+import AirshipFrameworkProxy
+
+@objc(AirshipPluginLoader)
+public class AirshipPluginLoader: NSObject, AirshipPluginLoaderProtocol {
+    public static func onApplicationDidFinishLaunching(
+        launchOptions: [UIApplication.LaunchOptionsKey : Any]?
+    ) {
+        AirshipAutopilot.shared.onLoad(launchOptions: launchOptions)
+    }
+}